simplify structure of ts-utils
