+++ src/alexlibs/ts-utils/Enum.ts
