+++ src/alexlibs/ts-utils/FnSwitch.ts
