+++ src/alexlibs/ts-utils/MapPromise.ts
