--- src/alexlibs/ts-utils/delay/Delay.ts
@@ -1,4 +1,0 @@
-export const delay =
-  <T>(ms: number) =>
-  (t: T): Promise<T> =>
-    new Promise(resolve => setTimeout(() => resolve(t), ms))
