+++ src/alexlibs/ts-utils/index.ts
@@ -1,10 +1,38 @@
-export * from './common/Common'
-export * from './common/CommonType'
-export * from './delay/Delay'
-export * from './duration/Duration'
-export * from './enum/Enum'
+export * from './Duration'
+export * from './Enum'
 export * from './env/Env'
 export * from './env/EnvParser'
-export * from './fnSwitch/FnSwitch'
-export * from './map/Map'
-export * from './mapPromise/MapPromise'
+export * from './FnSwitch'
+export * from './MapPromise'
+
+export const delay =
+  <T>(ms: number) =>
+  (t: T): Promise<T> =>
+    new Promise(resolve => setTimeout(() => resolve(t), ms))
+
+export const map = <T, R>(t: T | undefined, fn: (t: T) => R): R | undefined => {
+  return t !== undefined ? fn(t) : undefined
+}
+
+/**
+ * Functional for loop.
+ * mapFor(n, callback)
+ * is equivalent to
+ * [...new Array(n)].map((_, i) => callback(i))
+ * with better performance
+ */
+export const mapFor = <T>(n: number, callback: (i: number) => T): T[] => {
+  const result: T[] = new Array(n)
+  for (let i = 0; i < n; i++) {
+    result[i] = callback(i)
+  }
+  return result
+}
+
+export type DeepPartial<T> = {
+  [P in keyof T]?: T[P] extends Array<infer U>
+    ? Array<DeepPartial<U>>
+    : T[P] extends ReadonlyArray<infer U>
+    ? ReadonlyArray<DeepPartial<U>>
+    : DeepPartial<T[P]>
+}
