--- src/alexlibs/ts-utils/map/Map.ts
@@ -1,3 +1,0 @@
-export const map = <T, R>(t: T | undefined, fn: (t: T) => R): R | undefined => {
-  return t !== undefined ? fn(t) : undefined
-}
